+++ src/main/java/com/bitcamp/sc/domain/pay/service/impl/type/KakaoPay.java
@@ -37,8 +37,8 @@ public class KakaoPay {
 		
 		MultiValueMap<String, String> params = new LinkedMultiValueMap<String, String>();
 		params.add("cid", "TC0ONETIME");
-        params.add("partner_order_id", Integer.toString(orderInfo.getIdx()));
-        params.add("partner_user_id", Integer.toString(orderInfo.getMemberIdx()));
+        params.add("partner_order_id", Long.toString(orderInfo.getIdx()));
+        params.add("partner_user_id", Long.toString(orderInfo.getMemberIdx()));
         params.add("item_name", orderInfo.getCategory());
         params.add("quantity", "1");
         params.add("total_amount", Integer.toString(orderInfo.getPrice()));
@@ -74,8 +74,8 @@ public class KakaoPay {
 		MultiValueMap<String, String> params = new LinkedMultiValueMap<String, String>();
 		params.add("cid", "TC0ONETIME");
 		params.add("tid", kakaoPayReady.getTid());
-        params.add("partner_order_id", Integer.toString(orderInfo.getIdx()));
-        params.add("partner_user_id", Integer.toString(orderInfo.getMemberIdx()));
+        params.add("partner_order_id", Long.toString(orderInfo.getIdx()));
+        params.add("partner_user_id", Long.toString(orderInfo.getMemberIdx()));
         params.add("pg_token", pg_token);
         params.add("total_amount", Integer.toString(orderInfo.getPrice()));
