+++ src/main/java/com/bitcamp/sc/domain/tour/domain/TourDto.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 public class TourDto {
 	// 클라이언트 요청에 맵핑될 커맨드 클래스 db 저장 목적 x
 
-	private int midx;
+	private long midx;
 	private String selectDate;
 	private int tourPeople;
 	private int price;
